Refactor: create new example 'loaders' that handle table operations, and make test scenarios interact with this
This means that tests shouldn't need to change as we add implementations to make things pass; only the example loaders
should need to be updated to make use of the new versioning functionality.
