--- examples/src/main/scala/com/gu/tableversions/examples/EndToEndExample.scala
@@ -1,19 +1,0 @@
-package com.gu.tableversions.examples
-
-object EndToEndExample {
-
-  def main(args: Array[String]): Unit = {
-
-    // Import an unversioned table
-
-    // Update some of the partitions
-
-    // Query the metastore history
-
-    // Read the old version
-
-    // Read the new version
-
-  }
-
-}
